// src/common/enums/rbac.enum.ts
export enum RoleName {
  SUPER_ADMIN = 'SUPER_ADMIN',
  BUILDING_ADMIN = 'BUILDING_ADMIN',
  READ_ONLY = 'READ_ONLY',
}

export enum Action {
  CREATE = 'CREATE',
  READ = 'READ',
  UPDATE = 'UPDATE',
  DELETE = 'DELETE',
}

export enum Resource {
  USER = 'USER',
  BUILDING = 'BUILDING',
  APARTMENT = 'APARTMENT',
  EXPENSE = 'EXPENSE',
  OIL_DELIVERY = 'OIL_DELIVERY',
  COMMON_CHARGE = 'COMMON_CHARGE',
  PAYMENT = 'PAYMENT',
  DOCUMENT = 'DOCUMENT',
  ANNOUNCEMENT = 'ANNOUNCEMENT',
}